backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pypdf==5.1.0
python-docx==1.1.2
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.6.1
